work-in-progress MPI REST API
