+++ src/platform/CMakeLists.txt
@@ -2,6 +2,8 @@
 # Licensed under the MIT License.
 
 set(PLATFORM_INC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/inc)
+set(MODULESMANAGER_INC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/modulesmanager/inc)
+set(ORCHESTRATOR_INC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/orchestrator/inc)
 
 add_subdirectory(modulesmanager)
 add_subdirectory(orchestrator)
