+++ src/platform/inc/Mpi.h
@@ -44,10 +44,8 @@ void MpiClose(MPI_HANDLE clientSession);
 
 void MpiFree(MPI_JSON_STRING payload);
 void MpiInitialize(void);
-void MpiApiInitialize(void);
 void MpiDoWork(void);
 void MpiShutdown(void);
-void MpiApiShutdown(void);
 
 #ifdef __cplusplus
 }
