+++ src/platform/orchestrator/src/MpiApi.cpp
@@ -12,6 +12,7 @@
 #include "CommonUtils.h"
 #include "MpiApi.h"
 #include "ModulesManager.h"
+#include "Logging.h"
 
 static const char* g_socketPrefix = "/run/osconfig";
 static const char* g_mpiSocket = "/run/osconfig/mpid.sock";
@@ -23,11 +24,100 @@ static const char* g_componentName = "ComponentName";
 static const char* g_objectName = "ObjectName";
 static const char* g_payload = "Payload";
 
-static std::map<std::string, MPI_HANDLE> g_sessions;
+#define PLATFORM_LOGFILE "/var/log/osconfig_platform.log"
+#define PLATFORM_ROLLEDLOGFILE "/var/log/osconfig_platform.bak"
 
+static std::map<std::string, MPI_HANDLE> g_sessions;
 static Server server;
 
+class PlatformLog
+{
+public:
+    static OSCONFIG_LOG_HANDLE Get()
+    {
+        return m_log;
+    }
+
+    static void OpenLog()
+    {
+        m_log = ::OpenLog(PLATFORM_LOGFILE, PLATFORM_ROLLEDLOGFILE);
+    }
+
+    static void CloseLog()
+    {
+        ::CloseLog(&m_log);
+    }
+
+    static OSCONFIG_LOG_HANDLE m_log;
+};
+
 OSCONFIG_LOG_HANDLE PlatformLog::m_log = nullptr;
+
+namespace http
+{
+    const std::string CRLF = "\r\n";
+    const std::string contentLength = "Content-Length";
+    const std::string contentType = "Content-Type";
+    const std::string contentTypeJson = "application/json";
+
+    enum class Method
+    {
+        GET,
+        POST,
+        PUT,
+        DELETE,
+        PATCH,
+        INVALID
+    };
+
+    enum class StatusCode
+    {
+        OK = 200,
+        BAD_REQUEST = 400,
+        NOT_FOUND = 404,
+        INTERNAL_SERVER_ERROR = 500
+    };
+
+    struct Request
+    {
+        Method m_method;
+        std::string m_version;
+        std::string m_uri;
+        std::map<std::string, std::string> m_headers;
+        std::string m_body;
+
+        bool Parse(const std::string& request);
+    };
+
+    struct Response
+    {
+        StatusCode m_status;
+        std::map<std::string, std::string> m_headers;
+        std::string m_body;
+
+        std::string ToString();
+    };
+} // namespace http
+
+
+class Server
+{
+public:
+    Server() = default;
+    ~Server() = default;
+
+    void Listen();
+    static void Worker(Server& server);
+    void Stop();
+
+private:
+    int m_socketfd;
+    struct sockaddr_un m_addr;
+    socklen_t m_socketlen;
+
+    std::thread m_worker;
+    std::promise<void> m_exitSignal;
+};
 
 std::string CreateGUID()
 {
@@ -374,27 +464,6 @@ bool http::Request::Parse(const std::string& request)
                 }
             }
 
-            // Parse the body
-            // if (body.length() > 0)
-            // {
-            //     if (headers.find(http::contentType) != headers.end())
-            //     {
-            //         if (headers[http::contentType] == http::contentTypeJson)
-            //         {
-            //             rapidjson::Document document;
-            //             if (!document.Parse(body.c_str(), body.size()).HasParseError())
-            //             {
-            //                 body = document.GetString();
-            //             }
-            //             else
-            //             {
-            //                 result = false;
-            //             }
-            //         }
-            //     }
-            // }
-
-            // this->body = body;
             result = true;
         }
     }
